move legacy Table
