+++ src/components/legacy/Table/TableAmount.js
