+++ src/components/legacy/Table/package.json
