+++ src/screens/LiquidityPools/UnstakeTokensReview.js
@@ -27,7 +27,7 @@ import t from 'translations/translate';
 // components
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
 import TokenReviewSummary from 'components/ReviewSummary/TokenReviewSummary';
-import Table, { TableRow, TableLabel, TableAmount, TableTotal, TableFee } from 'components/Table';
+import Table, { TableRow, TableLabel, TableAmount, TableTotal, TableFee } from 'components/legacy/Table';
 import { Spacing } from 'components/legacy/Layout';
 import Button from 'components/legacy/Button';
 import Toast from 'components/Toast';
