+++ src/screens/SendCollectible/SendCollectibleConfirm.js
@@ -33,7 +33,7 @@ import { estimateTransactionAction, resetEstimateTransactionAction } from 'actio
 
 // components
 import ContainerWithHeader from 'components/legacy/Layout/ContainerWithHeader';
-import Table, { TableRow, TableLabel, TableAmount, TableTotal, TableUser, TableFee } from 'components/Table';
+import Table, { TableRow, TableLabel, TableAmount, TableTotal, TableUser, TableFee } from 'components/legacy/Table';
 import Button from 'components/legacy/Button';
 import { Spacing, ScrollWrapper } from 'components/legacy/Layout';
 import CollectibleReviewSummary from 'components/ReviewSummary/CollectibleReviewSummary';
