+++ src/screens/SendToken/SendTokenConfirm.js
@@ -29,7 +29,7 @@ import { CHAIN } from 'constants/chainConstants';
 // components
 import { Container, Content, Spacing } from 'components/layout/Layout';
 import HeaderBlock from 'components/HeaderBlock';
-import Table, { TableRow, TableLabel, TableAmount, TableTotal, TableUser, TableFee } from 'components/Table';
+import Table, { TableRow, TableLabel, TableAmount, TableTotal, TableUser, TableFee } from 'components/legacy/Table';
 import Button from 'components/legacy/Button';
 import TokenReviewSummary from 'components/ReviewSummary/TokenReviewSummary';
 import { BaseText } from 'components/legacy/Typography';
